+++ CHANGES.rst
@@ -454,9 +454,9 @@ Bug Fixes
 
   - Raise a `NotImplementedError` when fitting composite models. [#1915]
 
-  - Fixed bug in computation of ``Gaussian2D`` model. [#2038]
+  - Fixed bug in the computation of the ``Gaussian2D`` model. [#2038]
 
-  - Fixed bug in the computation of the `AiryDisk2D` model. [#2093]
+  - Fixed bug in the computation of the ``AiryDisk2D`` model. [#2093]
 
 - ``astropy.sphinx``